# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ProductFeedDto": {
				"description": "DTO pro zobrazení datového toku produktů\nObsahuje všechny properties včetně ID a timestamps",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"sourceSystem": {
						"type": "string"
					},
					"targetSystem": {
						"type": "string"
					},
					"format": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"refreshInterval": {
						"type": "number",
						"format": "double"
					},
					"transformationScript": {
						"type": "string"
					},
					"filterCriteria": {
						"type": "string"
					},
					"productIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"productCount": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"sourceSystem",
					"targetSystem",
					"format",
					"url",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProductFeedDto": {
				"description": "Data Transfer Object (DTO) pro vytvoření nového datového toku produktů",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"sourceSystem": {
						"type": "string"
					},
					"targetSystem": {
						"type": "string"
					},
					"format": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"refreshInterval": {
						"type": "number",
						"format": "double"
					},
					"transformationScript": {
						"type": "string"
					},
					"filterCriteria": {
						"type": "string"
					},
					"productIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"sourceSystem",
					"targetSystem",
					"format",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductFeedDto": {
				"description": "DTO pro aktualizaci existujícího datového toku produktů\nVšechny properties jsou volitelné",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"sourceSystem": {
						"type": "string"
					},
					"targetSystem": {
						"type": "string"
					},
					"format": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"refreshInterval": {
						"type": "number",
						"format": "double"
					},
					"transformationScript": {
						"type": "string"
					},
					"filterCriteria": {
						"type": "string"
					},
					"productIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ProductImageDto": {
				"description": "DTO pro zobrazení obrázku produktu\nObsahuje všechny properties obrázku produktu včetně ID a timestamps",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"altText": {
						"type": "string"
					},
					"sortOrder": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"imageType": {
						"type": "string"
					},
					"isMain": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"productId",
					"url",
					"sortOrder",
					"isMain",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DbErrorResponse": {
				"description": "Rozhraní pro odpověď s informací o chybě databáze",
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"recovered": {
						"type": "boolean"
					}
				},
				"required": [
					"success",
					"message",
					"error",
					"recovered"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "express-oracle-api",
		"version": "1.0.0",
		"description": "REST API s Express.js, TypeORM, Zod a Oracle",
		"contact": {}
	},
	"paths": {
...
		"/api/productfeeds": {
			"get": {
				"operationId": "GetProductFeeds",
				"responses": {
					"200": {
						"description": "Seznam datových toků produktů s paginací",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalPages": {
											"type": "number",
											"format": "double"
										},
										"limit": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/ProductFeedDto"
											},
											"type": "array"
										}
									},
									"required": [
										"totalPages",
										"limit",
										"page",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Získání všech datových toků produktů",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [
					{
						"description": "Filtrování podle aktivního stavu",
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Filtrování podle formátu",
						"in": "query",
						"name": "format",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProductFeed",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductFeedDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Vytvoření nového datového toku",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Data pro vytvoření datového toku",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductFeedDto",
								"description": "Data pro vytvoření datového toku"
							}
						}
					}
				}
			}
		},
		"/api/productfeeds/{id}": {
			"get": {
				"operationId": "GetProductFeed",
				"responses": {
					"200": {
						"description": "Datový tok nebo chybovou odpověď",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductFeedDto"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Získání konkrétního datového toku podle ID",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID datového toku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProductFeed",
				"responses": {
					"200": {
						"description": "Aktualizovaný datový tok nebo chybovou odpověď",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductFeedDto"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Aktualizace existujícího datového toku",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID datového toku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Data pro aktualizaci datového toku",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductFeedDto",
								"description": "Data pro aktualizaci datového toku"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProductFeed",
				"responses": {
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Odstranění datového toku",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID datového toku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/productfeeds/{id}/products": {
			"get": {
				"operationId": "GetProductsForFeed",
				"responses": {
					"200": {
						"description": "Seznam produktů napojených na datový tok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Získání produktů pro konkrétní datový tok",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID datového toku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "AddProductsToFeed",
				"responses": {
					"200": {
						"description": "Výsledek operace",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Přidání produktů k datovému toku",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID datového toku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Pole ID produktů k přidání",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"productIds": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"productIds"
								],
								"type": "object",
								"description": "Pole ID produktů k přidání"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateProductsInFeed",
				"responses": {
					"200": {
						"description": "Výsledek operace",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Aktualizace produktů v datovém toku (nahrazení stávajících)",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID datového toku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Pole ID produktů",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"productIds": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"productIds"
								],
								"type": "object",
								"description": "Pole ID produktů"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveProductsFromFeed",
				"responses": {
					"200": {
						"description": "Výsledek operace",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Odstranění produktů z datového toku",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID datového toku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Pole ID produktů k odstranění",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"productIds": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"productIds"
								],
								"type": "object",
								"description": "Pole ID produktů k odstranění"
							}
						}
					}
				}
			}
		},
		"/api/productfeeds/seed": {
			"post": {
				"operationId": "SeedProductFeeds",
				"responses": {
					"200": {
						"description": "Výsledek seedování",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Server error"
					}
				},
				"description": "Seedování testovacích dat datových toků",
				"tags": [
					"Interfaces - ProductFeeds"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/product-images/product/{productId}": {
			"get": {
				"operationId": "GetProductImagesByProductId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductImageDto"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při načítání obrázků produktu",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Získá všechny obrázky pro konkrétní produkt",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID produktu",
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/product-images/{id}/main/product/{productId}": {
			"put": {
				"operationId": "SetMainImage",
				"responses": {
					"200": {
						"description": "Obrázek byl nastaven jako hlavní",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Obrázek nebyl nalezen",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při nastavování hlavního obrázku",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Nastaví obrázek jako hlavní pro konkrétní produkt",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID obrázku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "ID produktu",
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/product-images/seed": {
			"post": {
				"operationId": "SeedTestData",
				"responses": {
					"201": {
						"description": "Testovací data úspěšně seedována",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při seedování testovacích dat",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Seedování testovacích dat obrázků produktů",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}